Time: 0 ms (100%), Space: 15.2 MB (30.14%) - LeetHub

+++ 2714-left-and-right-sum-differences/2714-left-and-right-sum-differences.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     vector<int> leftRightDifference(vector<int>& nums) {
-        short numsSize = nums.size();
+        int numsSize = nums.size();
 
         vector<int> leftSum, rightSum;
         leftSum.reserve(numsSize);
@@ -10,7 +10,7 @@ public:
         rightSum.reserve(numsSize);
         rightSum.push_back(0);
 
-        for (short i = 1; i < numsSize; ++i)
+        for (int i = 1; i < numsSize; ++i)
         {
             leftSum.push_back(leftSum[i - 1] + nums[i - 1]);
             rightSum.push_back(rightSum[i - 1] + nums[numsSize - i]);
